refactor(preprocessor): change preprocess code from `lib.rs` to `preprocess.rs`
